+++ docplex/cp/catalog.py
@@ -574,10 +574,14 @@ Oper_less_or_equal               = CpoOperation("lessOrEqual", "less_or_equal", 
 Oper_lexicographic               = CpoOperation("lexicographic", "lexicographic", None, -1, ( CpoSignature(Type_BoolExpr, (Type_IntExprArray, Type_IntExprArray)),) )
 Oper_strict_lexicographic        = CpoOperation("strictLexicographic", "strict_lexicographic", None, -1, ( CpoSignature(Type_BoolExpr, (Type_IntExprArray, Type_IntExprArray)),) )
 Oper_log                         = CpoOperation("log", "log", None, -1, ( CpoSignature(Type_FloatExpr, (Type_FloatExpr,)),) )
-Oper_logical_and                 = CpoOperation("and", "logical_and", "&&", 7, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr, Type_BoolExpr)),
+Oper_logical_and                 = CpoOperation("&&", "logical_and", "&&", 7, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr, Type_BoolExpr)),
+                                                                                 CpoSignature(Type_BoolExpr, (Type_BoolExprArray,)),) )
+Oper_logical_and_func            = CpoOperation("and", "and", None, -1, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr, Type_BoolExpr)),
                                                                                  CpoSignature(Type_BoolExpr, (Type_BoolExprArray,)),) )
 Oper_logical_not                 = CpoOperation("not", "logical_not", "!", 1, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr,)),) )
-Oper_logical_or                  = CpoOperation("or", "logical_or", "||", 8, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr, Type_BoolExpr)),
+Oper_logical_or                  = CpoOperation("||", "logical_or", "||", 8, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr, Type_BoolExpr)),
+                                                                               CpoSignature(Type_BoolExpr, (Type_BoolExprArray,)),))
+Oper_logical_or_func             = CpoOperation("or", "or", None, -1, ( CpoSignature(Type_BoolExpr, (Type_BoolExpr, Type_BoolExpr)),
                                                                                CpoSignature(Type_BoolExpr, (Type_BoolExprArray,)),))
 Oper_max                         = CpoOperation("max", "max", None, -1, ( CpoSignature(Type_IntExpr, (Type_IntExpr, Type_IntExpr)),
                                                                           CpoSignature(Type_IntExpr, (Type_IntExprArray,)),